requirements.txt
fastapi==0.115.2
uvicorn==0.30.1

sqlalchemy==2.0.31
pymysql==1.1.1

python-multipart==0.0.9
pydantic==2.9.2
requests==2.32.3
colorama==0.4.6

msal==1.31.0
python-dotenv==1.0.1
reportlab==4.2.2

# Beveiliging en authenticatie
python-jose[cryptography]==3.3.0
bcrypt==3.2.2
passlib==1.7.4
